fix: rename feature-toggle-service-v2.ts to feature-toggle-service.ts
