+++ src/lib/routes/client-api/feature.ts
@@ -3,7 +3,7 @@ import { Response } from 'express';
 import Controller from '../controller';
 import { IUnleashServices } from '../../types/services';
 import { IUnleashConfig } from '../../types/option';
-import FeatureToggleServiceV2 from '../../services/feature-toggle-service-v2';
+import FeatureToggleServiceV2 from '../../services/feature-toggle-service';
 import { Logger } from '../../logger';
 import { querySchema } from '../../schema/feature-schema';
 import { IFeatureToggleQuery } from '../../types/model';
