+++ src/lib/types/services.ts
@@ -18,7 +18,7 @@ import HealthService from '../services/health-service';
 import SettingService from '../services/setting-service';
 import SessionService from '../services/session-service';
 import UserFeedbackService from '../services/user-feedback-service';
-import FeatureToggleServiceV2 from '../services/feature-toggle-service-v2';
+import FeatureToggleService from '../services/feature-toggle-service';
 import EnvironmentService from '../services/environment-service';
 import FeatureTagService from '../services/feature-tag-service';
 import ProjectHealthService from '../services/project-health-service';
@@ -35,7 +35,8 @@ export interface IUnleashServices {
     environmentService: EnvironmentService;
     eventService: EventService;
     featureTagService: FeatureTagService;
-    featureToggleServiceV2: FeatureToggleServiceV2;
+    featureToggleService: FeatureToggleService;
+    featureToggleServiceV2: FeatureToggleService; // deprecated
     featureTypeService: FeatureTypeService;
     healthService: HealthService;
     projectHealthService: ProjectHealthService;
