+++ src/test/e2e/services/project-health-service.e2e.test.ts
@@ -1,6 +1,6 @@
 import dbInit, { ITestDb } from '../helpers/database-init';
 import getLogger from '../../fixtures/no-logger';
-import FeatureToggleServiceV2 from '../../../lib/services/feature-toggle-service-v2';
+import FeatureToggleServiceV2 from '../../../lib/services/feature-toggle-service';
 import { AccessService } from '../../../lib/services/access-service';
 import ProjectService from '../../../lib/services/project-service';
 import ProjectHealthService from '../../../lib/services/project-health-service';
